add modules for deletable and findable and remove shared methods from item and merchant classes
