+++ modules/deletable.rb
@@ -1,0 +1,7 @@
+module Deletable
+
+  def delete(id)
+    @all.delete(find_by_id(id))
+  end
+
+end
